12-17/1/2023

+++ lib/app/models/order_model.dart
@@ -4,7 +4,7 @@ class Order {
   String? orderId;
   String? status;
   String? currency;
-  int? amount;
+  double? amount;
   String? timeSlotId;
   static const String _orderId = 'orderId';
   static const String _status = 'status';
@@ -12,6 +12,7 @@ class Order {
   static const String _amount = 'amount';
   static const String _timeSlotId = 'timeSlotId';
   factory Order.fromMap(Map<String, dynamic> map) {
+
     return Order(
         orderId: map[_orderId],
         status: map[_status],
